move the logging method around, improve clarity for the logs, and add TODOs where action refactoring can happen in the near future.
