+++ src/sentry/workflow_engine/tasks/workflows.py
@@ -1,6 +1,5 @@
 import random
 import time
-from dataclasses import asdict
 from datetime import UTC, datetime
 from typing import Any
 
@@ -25,39 +24,10 @@ from sentry.workflow_engine.tasks.utils import (
     EventNotFoundError,
     build_workflow_event_data_from_event,
 )
-from sentry.workflow_engine.types import WorkflowEventData, WorkflowNotProcessable
+from sentry.workflow_engine.types import WorkflowEventData
 from sentry.workflow_engine.utils import log_context, scopedstats
 
 logger = log_context.get_logger(__name__)
-
-
-# TODO - this seems like it should live in processors.workflow,
-# Then we can create a `process_new_issue` or `process_event` kind
-# of method. That will wrap a new `process_actions_with_logs`
-# (and `process_actions`) method. Allowing us to have a much cleaner
-# interface here, and between those processor methods.
-def process_workflows_with_logs(
-    batch_client: DelayedWorkflowClient,
-    event_data: WorkflowEventData,
-    event_start_time: datetime,
-    detector: Detector | None = None,
-) -> None:
-    """
-    This method is used to create improved logging for `process_workflows`,
-    where we can capture the `WorkflowNotProcessable` results, and log them.
-
-    This should create a log for each issue, so we can determine why a workflow
-    did or did not trigger.
-    """
-    from sentry.workflow_engine.processors.workflow import process_workflows
-
-    evaluation = process_workflows(batch_client, event_data, event_start_time, detector)
-
-    if isinstance(evaluation, WorkflowNotProcessable):
-        logger.info("process_workflows.evaluation", extra=asdict(evaluation))
-    else:
-        # TODO - Log the triggered workflows
-        pass
 
 
 @instrumented_task(
@@ -75,6 +45,7 @@ def process_workflow_activity(activity_id: int, group_id: int, detector_id: int)
     The task will get the Activity from the database, create a WorkflowEventData object,
     and then process the data in `process_workflows`.
     """
+    from sentry.workflow_engine.processors.workflow import process_workflows_with_logs
 
     with transaction.atomic(router.db_for_write(Detector)):
         try:
@@ -133,6 +104,8 @@ def process_workflows_event(
     start_timestamp_seconds: float | None = None,
     **kwargs: dict[str, Any],
 ) -> None:
+    from sentry.workflow_engine.processors.workflow import process_workflows_with_logs
+
     recorder = scopedstats.Recorder()
     start_time = time.time()
